visionqa/requirements.txt
# VisionQA - Production Requirements
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0

# Google AI & Cloud
google-genai>=1.0.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
google-cloud-speech>=2.26.0
google-cloud-texttospeech>=2.16.0

# Browser Automation
playwright>=1.40.0

# WebSocket
websockets>=12.0

# HTTP Client
httpx>=0.27.0

# Utilities
python-multipart>=0.0.9
aiofiles>=23.2.0

